// src/modules/users/user-roles.enum.ts
export enum UserRole {
  ADMIN = 'ADMIN', // Usuário com permissões gerais para o bom funcionamento do sistema. (Integrante do projeto Special)
  SUPERVISOR = 'SUPERVISOR', // Usuário responsável pela coordenação do projeto AEE na Escola.
  TEACHER = 'TEACHER', // Professores integrantes do projeto AEE na Escola.
}